backend/requirements.txt
asgiref==3.6.0
Django==4.1.4
djangorestframework==3.14.0
pytz==2022.7
sqlparse==0.4.3
psycopg2-binary==2.9.5
cloudinary==1.30.0
dj3-cloudinary-storage==0.0.6
django-mdeditor==0.1.20